pyproject.toml
[tool.poetry]
name = "javelin-video-analysis"
version = "0.1.0"
description = "A video analysis program for javelin tracking, speed visualization, and acceleration heatmaps."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.5.3"
mediapipe = "^0.8.6"
numpy = "^1.21.0"
pyyaml = "^5.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"